pyproject.toml
[tool.poetry]
name = "discord-openai"
version = "0.1.0"
description = ""
authors = ["Jarigyani <coffee.mill.jp@gmail.com>"]
readme = "README.md"
packages = [{include = "discord_openai"}]

[tool.poetry.dependencies]
python = "^3.11.2"
py-cord = "^2.4.0"
openai = "^0.27.2"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
